Rebuild layout without grid.
As lovely as grid is, (a) I don't need it for this layout -- it's
quite traditional, after all! -- and (b) it had some *quirks* as far
as rendering goes.

Replace it with a simple `position: fixed` sidebar on large screens
and smart use of `max-width`.

Additionally, restructure so that the layout module contains *only*
layout, with the component modules actually describing the internals
and styles of those components.