Standardize and simplify "fancy block" style.
While the effect of pushing the blocks to the very edge for small
screens was *cool*, it had usability problems in terms of not being
clear what was happening when code overflowed the margins. It looked
slightly broken. Simplifying it so that the format expands once wide
enough (a) fixes that problem and (b) makes it easier to be consistent
across other "fancy block" styles like blockquotes.