Add new site dir with Netlify `_redirects`.
The `site` directory will be the home of *all* site structure going
forward, so that the root of the repository isn't cluttered with
templates and content and so on.

The redirects file here accounts for the Netlify subdomain and for
ensuring redirects to HTTPS from HTTP (though that should *also* be
handled by Cloudflare).