Add photography section to the site.
Use [today's post] from the *current* version of my site as some dummy
content while building this out. This prompts a question: how do I
want this section of the site to work? If I want it to be a photoblog,
I need to design a design that supports both posts like that one or
even long-form essays... or I need to move those to the writing
section of the site and keep this part *only* for photoblogging. Maybe
just links in the midst of the photoblog is fine, though!

[today's post]: https://v4.chriskrycho.com/2019/photography-ing-again.html