Start the work to extract WIP block nicely
This is a presentational step back in certain spots, but highlights
that I had slightly over-coupled some of the presentational elements
to their location (e.g. `aside`, `.note`, etc.). None of these should
actually be coupled to living inside an `article`, so I here stub out
a `_note.scss` module where `.note` will live, and so on; I will use
SCSS to implement the shared functionality and reuse it across each of
those UI elements, while making it work both inside and outside the
context of an `article`, so that they can be used independently as
makes sense.