Generalize `article` -> `.content-block`
In order to support displaying content (e.g. "note" items) in contexts
other than the main article context, and to properly decouple the
visuals of normal posts from the semantics of an article item, extract
all the elements that were previously part of `article` into a new
`.content-block` class definition. Apply it to the WIP macro so that the
WIP note renders correctly.