Move `link[rel="me"]` definitions to config.js
The `base.njk` template can now simply run a for loop to generate the
links, and editing them is simply a matter of adding or removing a
line from the JS object which exports them. Handily, this means they
are also available to code which imports the config—and are part of
the type of the config for that code.