Handle difference between `Page` and `Item`
Eleventy provides a `page` item, represented here by the `Page` type,
which does *not* have any of the associated items. It *also* presents
`Collection#items`, an array of items which are identical to `Page`s
except that they also have `templateContent` and `data` from the
item's metadata. Capture that here, in order to unblock work around
social graph data (#35).