Set correct ID for feed items.
The previous implementation was just ending up using the item's relative
URL because I wasn't passing in the `config` item when invoking the
`atom` macro. This change also cleans up the feed rendering along the
way, which will make it present more nicely in MailChimp (or any other
way I choose to forward this in the future).