Support book items without reviews
My 'Library' section should support not only *reviews* but also
ongoing commentary *as* I read. In that direction, make the 'review'
element of a 'book' field an optional hash including 'summary' and
'rating' instead of top-level values within 'book', and update all
references to 'review' and 'rating' to be 'review.summary' and
'review.rating' instead.