Use root collections in `toCollection`
In support of using nested directories for *internal organization*, not
*external* structure (as in the [`library/Eccentric Existence`][ee-dir]
directory created in f60e991a), rework the `toCollection` helper to use
the first directory *after* `./site` in the path.

[ee-dir]: https://github.com/chriskrycho/v5.chriskrycho.com/blob/f60e991a/site/library/Eccentric%20Existence

(A future refactor that would be nice here: making this work on the item
instead of on its slug; that way an override could be specified if I
wanted to have nested structures. I will do that *if and only if* I ever
actually want it, though.)

Resolves #81