Notes: finish the new section (MVP-style)
This doesn't actually cover everything I want or need from this:

- The styles for permalinks are not great; they should probably be
  integrated into the grid for the date in some way.
- The display on small screens honestly just ends up looking *bad*. D:
- The way I am excluding notes from the topics and archive pages *works*
  but if I'm going to have *everything* on the site, I might as well
  have *everything* on the site. I don't yet have a good design in place
  for that, though.
- This display format does *not* scale well without some kind of
  pagination. It may make sense for the notes archive to paginate the
  items on a monthly basis.

I will circle back to these items in the future; for now I'm just happy
to make this possible for myself. A couple quick Shortcuts on iOS/iPadOS
from now and I'll be able to post these *very* easily.