Style: profoundly simplify 'fancy blocks'
While I like the idea I had here in principle, in practice it did not
really pan out, and was a pretty persistent source of visual bugs. This
simplification still looks very nice, and in fact makes the horizontal
rhythm of individual posts much more consistent -- a genuine improvement
over the previous. Experiments are good; this one was a failure.