Layout: only show book view on reviews
While it's nice that in principle I *can* show the book cover on
every single post about a book... I don't actually *need* to and
it gets in the way of the content. It might be nice to integrate it
into the post meta in some way, but I can do that later!