Make all pages have useful metadata
This fixes a quirky little bug in 'draft' pages, which are present on
the website but not exposed in the feed or linked in any way (useful for
'drafts' today even if not optimal), caused by a confluence of the two
following bits:

-   the `page` item 11ty supplies doesn't actually provide enough data
    to be useful for e.g. metadata generation at the point where you're
    rendering a given page, so it has to be run through a filter which
    gets the relevant data out of the full set of data for the page via
    the version of the page available in the collection view

-   I was accounting for that and filtering, but when I started
    filtering all content through the `live` collection, I accidentally
    included *this* in that change: I implemented that by doing a naive
    find-and-replace operation

This switches the per-item lookup to use `collections.all` instead of
`collections.live` so that all posts, whether live or not, get their
metadata rendered correctly. Yeeeeeesh!