Photos: add a title to the latest
I need to special-case the handling for micro.blog, but that should be
just that: a special case for that feed, *not* something I do for my
photo posts in general. That's for a number of reasons, not least that
I don't want full-bleed images showing up in the archive views on my
site! They're not like 'notes' in that regard.