Support 'ghost' collections
This allows me to use folders as 'collections' by default, while opting
out for organizational-only collections.

(Thinking about `lx` design: should this be the default? Or should it
be the other way: folders only become 'collections' when explicitly
opting in? Worth playing with both ways, but given that I only need to
do it in a couple places in the current design, opt-*out* seems like
it's probably a reasonable starting point.)