Switch over a bunch of MMD abbreviations to HTML
This will be an ongoing-but-occasional task. The MMD abbreviation format
is convenient, but it’s also slow and not *that* useful, and it requires
a multi-pass parse, which I want to move my content away from so that I
can switch to something like pulldown-cmark as part of lx... eventually.