Fix typo in stateful/class-backed example
I'm pretty sure that `this.value` is a typo, the only value in that example is the one that gets extracted from the `input` event `event.target.value` and I doubt it has anything to do with the button.

BTW I use to use `{{not this.name}}`, which, given that `name` is string (or string | null | undefined) type, should be the same thing. Are you checking the length instead for clarity purposes?