Start collapsing some topics
I realized (while chatting with @pzuraq) something I'd been thinking
about for a while: I need to publish as few topics as possible to keep
my site generation smaller. (Someday maybe rewriting it in Rust will be
great, but for now I need to deal with the constraints I have.) This is
the first step in that: finding actual duplicate topics, or topics
which duplicate other infrastructure I've since built (in particular,
being able to group posts for books, series, etc.) and collapsing them
together. This commit is the first step *in* that first step: doing
that for just a handful of topics.