Pre-parse YAML content into Markdown
This is about 1.2× faster than the previous version, which repeatedly
invoked the `inlineMd` helper for each of the fields handled here; that
added up to *thousands* of invocations across the files. It's still a
4.8s build, and I would still like that to be much faster, but that
will not happen with 11ty's mediocre data structures and algorithms
(case in point: why is there literally *no* documentation of doing this
thing anywhere? Because no one thinks about or cares to do it!). To do
that, I will need to build my own thing (and maybe, just maybe, do it
in something which is not JS!).