Work around missing 11ty data cascade API
11ty does not seem to provide a good hook for preprocessing the resolved
data in “the cascade”. That means anything I was putting in a shared
`.11tydata.json` file was not getting preprocessed. The main place I
know this is an issue is the 2022 in Review series, though very likely I
will discover others over time. For now, I have fixed that one and left
myself a note for the future. If I find an *easy* fix, I will use it,
but as with other recent changes, the goal is actually to just get the
heck *off* of 11ty.