Journal: Create a new 'Read the Code' collection
This will be a nice parallel to 'Read the Manual', and the first item in
it will be 'Using Drop for Safety in Rust'.

(This continues to suggest to me that I may want to do something careful
or clever or both such that v6 can easily inherit a lot of this content,
but I do not yet know what or how I will approach that? Lots of sticking
`link rel=canonical` on posts I port over, maybe? Just a bucket of links
which go back to v5? Something to keep mulling on.)