Rename tag: 'ratchets' -> 'software development'
This is a slightly more interesting one than most: I think a “ratchets”
topic *could* be interesting, but right now there is only one entry in
it. I am removing it for the moment, but I could also, in the future, go
back and *add* it to a bunch of posts where it fits.