spike: try to switch to native Node type stripping
There are bugs with the acorn integration, and I seem to be among the
first people to have tried this (at least: there are no issues about
it). It *builds* just fine… but it isn’t any *faster* than it was before
so I don’t see any point to using it this way. I could pull some of this
forward for the sake of being up to date, but: it doesn’t matter, and I
am tracking fairly steadily toward being on lx instead. So I will leave
this in my history for posterity (including pushing a branch for it to
GitHub), but that’s it.